Refactor into base + put max kernels into pool
