+++ hotpot_km/async_mapping.py
@@ -1,12 +1,11 @@
 
 from jupyter_server.services.kernels.kernelmanager import AsyncMappingKernelManager
 
-from . import LimitedKernelManager, PooledKernelManager
+from .base import PooledKernelManager
 
 
 class AsyncLimitedPooledMappingKernelManager(
     AsyncMappingKernelManager,
-    PooledKernelManager,
-    LimitedKernelManager
+    PooledKernelManager
 ):
     pass
